requirements.txt
torch>=2.1.0
transformers>=4.49.0
accelerate>=0.33.0
datasets>=2.19.0
pillow>=10.0.0
tqdm>=4.66.0
pyyaml>=6.0.0
pandas>=2.0.0
qwen-vl-utils>=0.0.8
huggingface_hub>=0.23.0
